Rename `NewStorage` to `Storage`
...and `storage2` to `storage`
