+++ packages/tre/src/plugins/d1/gateway.ts
@@ -7,7 +7,7 @@ import type { Database as DatabaseType } from "better-sqlite3";
 import { z } from "zod";
 import { Response } from "../../http";
 import { HttpError, Log } from "../../shared";
-import { NewStorage } from "../../storage2";
+import { Storage } from "../../storage";
 import splitSqlQuery from "./splitter";
 
 export const D1ValueSchema = z.union([
@@ -153,7 +153,7 @@ interface ChangesLastRowResult {
 export class D1Gateway {
   private readonly db: DatabaseType;
 
-  constructor(private readonly log: Log, storage: NewStorage) {
+  constructor(private readonly log: Log, storage: Storage) {
     this.db = storage.db;
   }
