+++ packages/tre/src/storage/storage.ts
@@ -4,19 +4,18 @@ import Database from "better-sqlite3";
 import { BlobStore, FileBlobStore, MemoryBlobStore } from "./blob";
 import { TypedDatabase } from "./sql";
 
-// TODO: rename to `Storage` and move to `src/storage` once all gateways migrated
-export interface NewStorage {
+export interface Storage {
   db: TypedDatabase;
   blob: BlobStore;
 }
 
-export function createMemoryStorage(): NewStorage {
+export function createMemoryStorage(): Storage {
   const db = new Database(":memory:") as TypedDatabase;
   const blob = new MemoryBlobStore();
   return { db, blob };
 }
 
-export function createFileStorage(root: string): NewStorage {
+export function createFileStorage(root: string): Storage {
   root = path.resolve(root);
   fs.mkdirSync(root, { recursive: true });
   const db = new Database(path.join(root, "db.sqlite")) as TypedDatabase;
